# add runtime perf stats; remove some async