--- a/lib/src/generator/empty_generator.dart
+++ b/lib/src/generator/empty_generator.dart
@@ -28,6 +28,6 @@ class EmptyGenerator extends Generator {
   }
 }
 
-Future<Generator> initEmptyGenerator(DartdocOptionContext config) async {
+Generator initEmptyGenerator(DartdocOptionContext config) {
   return EmptyGenerator();
 }