--- a/lib/src/model/package_graph.dart
+++ b/lib/src/model/package_graph.dart
@@ -10,6 +10,8 @@ import 'package:analyzer/file_system/file_system.dart';
 import 'package:analyzer/src/generated/sdk.dart' show DartSdk, SdkLibrary;
 // ignore: implementation_imports
 import 'package:analyzer/src/generated/source.dart' show Source;
+// ignore: implementation_imports
+import 'package:analyzer/src/generated/timestamped_data.dart';
 import 'package:collection/collection.dart';
 import 'package:dartdoc/src/dartdoc_options.dart';
 import 'package:dartdoc/src/failure.dart';