--- a/test/end2end/dartdoc_test.dart
+++ b/test/end2end/dartdoc_test.dart
@@ -57,7 +57,7 @@ void main() {
     late Folder tempDir;
 
     setUpAll(() async {
-      var optionSet = await DartdocOptionRoot.fromOptionGenerators(
+      var optionSet = DartdocOptionRoot.fromOptionGenerators(
           'dartdoc',
           [createDartdocProgramOptions, createLoggingOptions],
           pubPackageMetaProvider);