This was work done during a very casual look at the runtime perf of dartdoc; this PR does not contain any actual perf improvements.

- add some simple runtime perf stats, to how much time we're spending in various parts of doc generation
- remove some uses of async (where we could just collapse to simple sync calls, and then iteratively remove async from callers and APIs; this makes the perf a little easier to trace in things like DevTools
- fix a few minor issues in dartdoc's own API comments

Here's some sample perf stats (from documenting dartdoc itself):

```
generateDocs                     31,323ms
  buildPackageGraph              2,3106ms
    _calculatePackageMap             18ms
    getLibraries                 16,787ms
    initializePackageGraph        6,299ms
  generator.generate              1,865ms
  validateLinks                   6,343ms
```