@kevmoo 2022-02-08
There's a lint for await only Future, right?