@devoncarew 2022-02-08
> How did you track those down? I'd love to prevent it in the future. Can you think of a way to track unnecessary async statically? Maybe I can bake it into whole-world analysis.

I didn't do anything fancy - just remove async from simple methods and APIs, trying to fix up all the resulting static analysis issues, and backing a change out if it bottoms out somewhere that really does need to be async (package:package_config was one of those places - there are some APIs there where we could easily offer sync versions).